Commit re-routing
