--- app/controllers/api/v0/site_config_controller.rb
@@ -1,10 +1,0 @@
-module Api
-  module V0
-    class SiteConfigController < ApiController
-      before_action :authenticate_super_admin!
-      def show
-        @site_config = SiteConfig.all
-      end
-    end
-  end
-end
