+++ app/views/api/v0/admin/configs/show.json.jbuilder
